refactor: mv token list utils to lib
