+++ src/hooks/useENSAvatar.ts
@@ -1,7 +1,7 @@
 import { namehash } from '@ethersproject/hash'
+import uriToHttp from 'lib/utils/uriToHttp'
 import { useEffect, useMemo, useState } from 'react'
 import { safeNamehash } from 'utils/safeNamehash'
-import uriToHttp from 'utils/uriToHttp'
 
 import { useSingleCallResult } from '../state/multicall/hooks'
 import { isAddress } from '../utils'
