+++ src/lib/utils/fetchTokenList.ts
@@ -2,7 +2,7 @@ import { TokenList } from '@uniswap/token-lists'
 import { ValidateFunction } from 'ajv'
 
 import contenthashToUri from './contenthashToUri'
-import { parseENSAddress } from './parseENSAddress'
+import parseENSAddress from './parseENSAddress'
 import uriToHttp from './uriToHttp'
 
 // lazily get the validator the first time it is used
