+++ src/lib/utils/uriToHttp.ts
@@ -1,10 +1,12 @@
 /**
- * Given a URI that may be ipfs, ipns, http, or https protocol, return the fetch-able http(s) URLs for the same content
+ * Given a URI that may be ipfs, ipns, http, https, ar, or data protocol, return the fetch-able http(s) URLs for the same content
  * @param uri to convert to fetch-able http url
  */
 export default function uriToHttp(uri: string): string[] {
   const protocol = uri.split(':')[0].toLowerCase()
   switch (protocol) {
+    case 'data':
+      return [uri]
     case 'https':
       return [uri]
     case 'http':
@@ -15,6 +17,9 @@ export default function uriToHttp(uri: string): string[] {
     case 'ipns':
       const name = uri.match(/^ipns:(\/\/)?(.*)$/i)?.[2]
       return [`https://cloudflare-ipfs.com/ipns/${name}/`, `https://ipfs.io/ipns/${name}/`]
+    case 'ar':
+      const tx = uri.match(/^ar:(\/\/)?(.*)$/i)?.[2]
+      return [`https://arweave.net/${tx}`]
     default:
       return []
   }
